backend/requirements.txt
django
simplejson
pylint-django
coverage
networkx
pylab
scipy
numpy
matplotlib
markdown
djangorestframework
django-filter